step2/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
PyJWT==2.6.0
Werkzeug==2.3.3